+++ src/MooVC.Syntax.CSharp.Tests/Members/DirectiveTests/WhenValidateIsCalled.cs
@@ -1,8 +1,7 @@
-namespace MooVC.Syntax.CSharp.Containers.DirectiveTests;
+namespace MooVC.Syntax.CSharp.Members.DirectiveTests;
 
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using MooVC.Syntax.CSharp.Members;
 
 public sealed class WhenValidateIsCalled
 {
