+++ src/MooVC.Syntax.CSharp.Tests/MooVC.Syntax.CSharp.Tests.csproj
@@ -2,4 +2,7 @@
   <PropertyGroup>
     <TargetFrameworks Condition="'$(IsCodespace)' != 'true'">net8.0;net9.0;net10.0</TargetFrameworks>
   </PropertyGroup>
+  <ItemGroup>
+    <Folder Include="Containers\" />
+  </ItemGroup>
 </Project>
