+++ src/MooVC.Syntax.CSharp/Members/Directive.Resources.resx
