+++ src/MooVC.Syntax.CSharp/MooVC.Syntax.CSharp.csproj
@@ -40,7 +40,7 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Definition.Resources.resx</DependentUpon>
     </Compile>
-    <Compile Update="Containers\Directive.Resources.Designer.cs">
+    <Compile Update="Members\Directive.Resources.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
       <DependentUpon>Directive.Resources.resx</DependentUpon>
@@ -104,7 +104,7 @@
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Definition.Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
-    <EmbeddedResource Update="Containers\Directive.Resources.resx">
+    <EmbeddedResource Update="Members\Directive.Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Directive.Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
